pyproject.toml
[tool.poetry]
name = "condorgp"
version = "0.1.0"
description = "Genetic Programming fintech AI autotrading project"
authors = ["Hugh Harford <hugh.harford@poscoconsulting.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
pytz = "^2025.1"
pandas = "^2.3.3"
nautilus-trader = "^1.222.0"
time-uuid = "^0.2.0"
setuptools = ">=75"
wheel = ">=0.29"
twine = "^5.1.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.evolution.dependencies]
numpy = "2.4.1"
deap = "^1.4.1"

[tool.poetry.group.comms.dependencies]
fastapi = "^0.115.6"
uvicorn = "^0.19.0"
pika = "^1.3.2"

[tool.poetry.group.database.dependencies]
sqlalchemy = "^1.4.42"
psycopg2-binary = "^2.9.5"
alembic = "^1.14.0"
pydantic = "^2.10.6"

[tool.poetry.group.cluster.dependencies]
# kubernetes = "^32.0.0" # taken out so doesn't build on the worker, for now
net-tools = "^0.1.2"

[tool.poetry.group.cloud.dependencies]
google-api-core = "^2.20.0"
gcloud = "^0.18.3"
gsutil = "^5.30"
google-cloud-storage = "*"
google-cloud-datacatalog = "^3.20.1"
google-cloud-compute = "^1.19.2"

[tool.poetry.group.test.dependencies]
pytest = "<9.0.0" # 8.4.2 is the last <9 - were on 9.0.2
pytest-bdd = "^8.1.0"
coverage = "^7.6.4"
kubernetes = "^35.0.0"
pick = "^2.4.0"

[tool.poetry.group.dev.dependencies]
docker = "^7.1.0"
ipython = "^8.28.0"
memory-profiler = "^0.61.0"
line-profiler = "^4.1.3"
black = "^24.10.0"
flake8 = "^7.1.1"
yapf = "^0.40.2"
six = ">=1.14"
gh = "^0.0.4"

[tool.poetry.group.optional.dependencies]
python-on-whales = "^0.75.1"
jq = "^1.8.0"
joblib = "^1.4.2"
termcolor = "^2.5.0"
memoized-property = "^1.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
